+++ app/src/main/java/com/cityCatTarot/errors/CardNotFoundException.java
@@ -4,7 +4,7 @@ package com.cityCatTarot.errors;
  * 카드를 찾을 수 없는 예외.
  */
 public class CardNotFoundException extends RuntimeException {
-    public CardNotFoundException(int cardId) {
+    public CardNotFoundException(Long cardId) {
         super("Card not found: " + cardId);
     }
 }
